requirements/local.txt
-r base.text

django-debug-toolbar==0.11.0
